pyproject.toml
[project]
name = "kanban"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.5",
    "django-allauth>=65.3.1",
    "django-extensions>=3.2.3",
    "django-browser-reload>=1.17.0",
    "pillow>=11.1.0",
    "requests>=2.32.5",
    "jwt>=1.4.0",
    "django-cotton>=2.1.3",
    "djangorestframework>=3.16.1",
    "django-filter>=25.2",
    "ruff>=0.14.2",
]


[tool.ruff]
line-length = 500  # Strict PEP 8, or use 88 for Black compatibility
target-version = "py311"

exclude = [
    "migrations",
    "__pycache__",
    "venv",
    ".venv",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors (PEP 8)
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
]

ignore = []

[tool.ruff.lint.isort]
known-first-party = ["Kanban"]


[tool.djlint]
profile = "django"
ignore = "H031,H005,H030"  
indent = 2
max_line_length = 120

[dependency-groups]
dev = [
    "djlint>=1.36.4",
]
